:truck: Moving and renaming Task Notification files
Files moved and renamed for better understanding and making it more
generic for future usages.
