+++ app/src/main/java/com/escodro/alkaa/ui/task/alarm/TaskReceiver.kt
@@ -6,19 +6,22 @@ import android.content.Intent
 import androidx.work.Data
 import androidx.work.OneTimeWorkRequest
 import androidx.work.WorkManager
+import com.escodro.alkaa.ui.task.alarm.notification.TaskNotificationScheduler
+import com.escodro.alkaa.ui.task.alarm.worker.TaskNotifierWorker
+import com.escodro.alkaa.ui.task.alarm.worker.TaskReschedulerWorker
 import org.koin.standalone.KoinComponent
 import timber.log.Timber
 
 /**
  * [BroadcastReceiver] to be notified by the [android.app.AlarmManager].
  */
-class TaskAlarmReceiver : BroadcastReceiver(), KoinComponent {
+class TaskReceiver : BroadcastReceiver(), KoinComponent {
 
     override fun onReceive(context: Context?, intent: Intent?) {
         Timber.d("onReceive() - intent ${intent?.action}")
 
         when (intent?.action) {
-            TaskAlarmManager.ALARM_ACTION -> onAlarm(intent)
+            TaskNotificationScheduler.ALARM_ACTION -> onAlarm(intent)
             Intent.ACTION_BOOT_COMPLETED -> onBootCompleted()
         }
     }
@@ -33,11 +36,11 @@ class TaskAlarmReceiver : BroadcastReceiver(), KoinComponent {
     private fun onAlarm(intent: Intent?) {
         Timber.d("onAlarm")
 
-        val taskId = intent?.getLongExtra(TaskAlarmManager.EXTRA_TASK, 0) ?: return
+        val taskId = intent?.getLongExtra(TaskNotificationScheduler.EXTRA_TASK, 0) ?: return
         val data = Data.Builder()
-        data.putLong(TaskAlarmManager.EXTRA_TASK, taskId)
+        data.putLong(TaskNotificationScheduler.EXTRA_TASK, taskId)
 
-        val worker = OneTimeWorkRequest.Builder(TaskAlarmWorker::class.java)
+        val worker = OneTimeWorkRequest.Builder(TaskNotifierWorker::class.java)
         worker.setInputData(data.build())
 
         WorkManager.getInstance().enqueue(worker.build())
