+++ app/src/main/java/com/escodro/alkaa/ui/task/alarm/notification/TaskNotificationChannel.kt
@@ -1,4 +1,4 @@
-package com.escodro.alkaa.ui.task.notification
+package com.escodro.alkaa.ui.task.alarm.notification
 
 import android.app.NotificationChannel
 import android.app.NotificationManager
